Add support for $script_phase_pre_upload.
This is set to a script that is executed by the upload script prior to
starting the actual upload, and is useful for performing any sanity
checks on what is about to be uploaded.

Add variables used by Joyent to ensure a certain number of packages
built ($pkg_rsync_min) and that a list of core packages all built
successfully ($pkg_rsync_required_pkgs) to the example config.