lang: Disable all non-Joyent GCCs.
We don't ever want users to accidentally install GCC versions that do
not have the correct illumos patches applied, and this saves on wasted
bulk builds for the versions that don't build correctly anyway.

It is also required for versions that come from pkgsrc-joyent as pbulk
will incorrectly take whichever has the highest version, even if marked
BROKEN, so this ensures that our version is always used.