mk: Implement checksum script in awk.
The existing shell script runtime is quadratic with the number of
distfiles that need to be handled.  Ordinarily this is not a problem,
but with some packages, for example those that use Go modules, this can
quickly become a serious bottleneck.

Rewriting the script in awk takes, as an example, the runtime of "bmake
checksum" in www/grafana (during an update to 7.1.3) containing over 600
distfiles from:

  real    8m12.049s
  user    7m32.178s
  sys     0m37.830s

down to just:

  real    0m3.936s
  user    0m3.015s
  sys     0m0.759s

I've tried to preserve the output of the shell script, and support all
of its idiosyncrasies, for example "IGNORE" support.