gobject-introspection: Prune LD_LIBRARY_PATH on SunOS.
The default is to add every -L path to LD_LIBRARY_PATH which is set in
the environment when calling the C compiler via g-ir-scanner.  This
causes problems on SunOS if any of those paths contain libraries that
then take priority over the RUNPATH from within libraries we are linking
against.

For example when building pango, the cairo libraries picked up the wrong
libgcc_s.so.1 from the platform instead of from pkgsrc and broken the
build.

This was previously fixed in the pkgsrc package, but turned out to cause
problems on Darwin and I reverted without understanding the original
reasons for the change.  Keep this one locally for now until we can
ensure it will not cause regressions elsewhere.